+++ pdata/internal/cmd/pdatagen/internal/base_structs.go
@@ -70,7 +70,8 @@ func Test${structName}_CopyTo(t *testing.T) {
 	assert.EqualValues(t, orig, ms)
 	orig = generateTest${structName}()
 	orig.CopyTo(ms)
-	assert.EqualValues(t, orig, ms)`
+	assert.EqualValues(t, orig, ms)
+}`
 
 const messageValueGenerateTestTemplate = `func generateTest${structName}() ${structName} {
 	tv := New${structName}()
@@ -167,13 +168,6 @@ func (ms *messageValueStruct) generateTests(sb *strings.Builder) {
 			panic(name)
 		}
 	}))
-	sb.WriteString(newLine)
-	for _, f := range ms.fields {
-		if psf, ok := f.(*primitiveSliceField); ok {
-			psf.generateTestItemMutation(sb)
-		}
-	}
-	sb.WriteString("}")
 
 	// Write accessors tests for the struct
 	for _, f := range ms.fields {